% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_unicode.R
\name{check_unhandled}
\alias{check_unhandled}
\title{Internal function to check for unhandled unicode}
\usage{
check_unhandled(x)
}
\arguments{
\item{x}{Character vector}
}
\description{
Internal function to check for unhandled unicode
}
\keyword{internal}
