% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_harmonizer.R
\name{fetch_molecular_weight}
\alias{fetch_molecular_weight}
\title{Fetch molecular weight via ComptoxR API}
\usage{
fetch_molecular_weight(dtxsids)
}
\arguments{
\item{dtxsids}{Character vector of unique DTXSIDs}
}
\value{
List containing named numeric \code{values} and named logical
\code{lookup_failed} vectors.
}
\description{
Fetch molecular weight via ComptoxR API
}
\keyword{internal}
