% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{format_file_size}
\alias{format_file_size}
\title{Format file size for display}
\usage{
format_file_size(size_bytes)
}
\arguments{
\item{size_bytes}{Numeric. File size in bytes}
}
\value{
Character. Formatted file size
}
\description{
Format file size for display
}
