% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_harmonizer.R
\name{get_molarity_scale}
\alias{get_molarity_scale}
\title{Get molarity scale factor for conversion to mg/L}
\usage{
get_molarity_scale(unit)
}
\arguments{
\item{unit}{Character - the molarity unit}
}
\value{
Numeric scale factor
}
\description{
Formula: mg/L = molarity * molecular_weight * scale_factor
}
\keyword{internal}
