% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_harmonizer.R
\name{get_ppx_conversion_factor}
\alias{get_ppx_conversion_factor}
\title{Get conversion factor for ppb/ppm to target unit}
\usage{
get_ppx_conversion_factor(unit)
}
\arguments{
\item{unit}{Character - ppb or ppm}
}
\value{
Numeric conversion factor
}
\description{
Get conversion factor for ppb/ppm to target unit
}
\keyword{internal}
