% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_harmonizer.R
\name{get_unit_synonyms}
\alias{get_unit_synonyms}
\title{Load unit synonyms internally via system.file}
\usage{
get_unit_synonyms()
}
\value{
Tibble with synonym mappings or NULL if not found
}
\description{
Load unit synonyms internally via system.file
}
\keyword{internal}
