% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_detection.R
\name{handle_merged_cells}
\alias{handle_merged_cells}
\title{Handle merged cells by filling down first column}
\usage{
handle_merged_cells(df)
}
\arguments{
\item{df}{Data frame to process}
}
\value{
Data frame with filled merged cells
}
\description{
Handle merged cells by filling down first column
}
