% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clean_data.R
\name{mod_clean_data_server}
\alias{mod_clean_data_server}
\title{Clean Data Module - Server}
\usage{
mod_clean_data_server(id, data_store, on_cleaning_complete = NULL)
}
\arguments{
\item{id}{Module namespace ID}

\item{data_store}{Reactive values store from main app}

\item{on_cleaning_complete}{Callback function to execute after cleaning completes (for navigation)}
}
\value{
Reactive list with cleaning indicators
}
\description{
Clean Data Module - Server
}
