% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clean_data.R
\name{mod_clean_data_ui}
\alias{mod_clean_data_ui}
\title{Clean Data Module - UI}
\usage{
mod_clean_data_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for clean data tab
}
\description{
Clean Data Module - UI
}
