% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_preview.R
\name{mod_data_preview_server}
\alias{mod_data_preview_server}
\title{Data Preview Module - Server}
\usage{
mod_data_preview_server(id, data_store, preview_rows)
}
\arguments{
\item{id}{Module namespace ID}

\item{data_store}{Reactive values store from main app}

\item{preview_rows}{Reactive integer for number of rows to preview}
}
\value{
NULL (module has no return values)
}
\description{
Data Preview Module - Server
}
