% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_preview.R
\name{mod_data_preview_ui}
\alias{mod_data_preview_ui}
\title{Data Preview Module - UI}
\usage{
mod_data_preview_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for data preview tab
}
\description{
Data Preview Module - UI
}
