% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataset_context.R
\name{mod_dataset_context_server}
\alias{mod_dataset_context_server}
\title{Dataset Context Module - Server}
\usage{
mod_dataset_context_server(id, data_store)
}
\arguments{
\item{id}{Module namespace ID}

\item{data_store}{Reactive values store from main app.}
}
\value{
Reactive list with site context state.
}
\description{
Dataset Context Module - Server
}
