% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataset_context.R
\name{mod_dataset_context_ui}
\alias{mod_dataset_context_ui}
\title{Dataset Context Module - UI}
\usage{
mod_dataset_context_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for dataset context curation.
}
\description{
Dataset Context Module - UI
}
