% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_detection_info.R
\name{mod_detection_info_server}
\alias{mod_detection_info_server}
\title{Detection Info Module - Server}
\usage{
mod_detection_info_server(id, data_store)
}
\arguments{
\item{id}{Module namespace ID}

\item{data_store}{Reactive values store from main app}
}
\value{
NULL (module has no return values)
}
\description{
Detection Info Module - Server
}
