% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_detection_info.R
\name{mod_detection_info_ui}
\alias{mod_detection_info_ui}
\title{Detection Info Module - UI}
\usage{
mod_detection_info_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for detection info tab
}
\description{
Detection Info Module - UI
}
