% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_upload.R
\name{mod_file_upload_ui}
\alias{mod_file_upload_ui}
\title{File Upload Module - UI}
\usage{
mod_file_upload_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
tagList of sidebar UI elements
}
\description{
File Upload Module - UI
}
