% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_harmonize.R
\name{mod_harmonize_server}
\alias{mod_harmonize_server}
\title{Harmonize Module - Server}
\usage{
mod_harmonize_server(id, data_store)
}
\arguments{
\item{id}{Module namespace ID}

\item{data_store}{Shared reactive data store}
}
\value{
NULL (module used for side effects on data_store)
}
\description{
Harmonize Module - Server
}
