% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_harmonize.R
\name{mod_harmonize_ui}
\alias{mod_harmonize_ui}
\title{Harmonize Module - UI}
\usage{
mod_harmonize_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for harmonize tab
}
\description{
Harmonize Module - UI
}
