% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_raw_data.R
\name{mod_raw_data_server}
\alias{mod_raw_data_server}
\title{Raw Data Module - Server}
\usage{
mod_raw_data_server(id, data_store)
}
\arguments{
\item{id}{Module namespace ID}

\item{data_store}{Reactive values store from main app}
}
\value{
NULL (module has no return values)
}
\description{
Raw Data Module - Server
}
