% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_raw_data.R
\name{mod_raw_data_ui}
\alias{mod_raw_data_ui}
\title{Raw Data Module - UI}
\usage{
mod_raw_data_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for raw data tab
}
\description{
Raw Data Module - UI
}
