% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_results.R
\name{mod_review_results_server}
\alias{mod_review_results_server}
\title{Review Results Module - Server}
\usage{
mod_review_results_server(id, data_store)
}
\arguments{
\item{id}{Module namespace ID}

\item{data_store}{Reactive values store from main app}
}
\value{
NULL (module has no return values)
}
\description{
Review Results Module - Server
}
