% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_results.R
\name{mod_review_results_ui}
\alias{mod_review_results_ui}
\title{Review Results Module - UI}
\usage{
mod_review_results_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for review results tab
}
\description{
Review Results Module - UI
}
