% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_run_curation.R
\name{mod_run_curation_ui}
\alias{mod_run_curation_ui}
\title{Run Curation Module - UI}
\usage{
mod_run_curation_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for run curation tab
}
\description{
Run Curation Module - UI
}
