% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tag_columns.R
\name{mod_tag_columns_ui}
\alias{mod_tag_columns_ui}
\title{Tag Columns Module - UI}
\usage{
mod_tag_columns_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
UI elements for tag columns tab
}
\description{
Tag Columns Module - UI
}
