% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_reference.R
\name{reference_list_help_text}
\alias{reference_list_help_text}
\title{Reference-list help text}
\usage{
reference_list_help_text(type)
}
\arguments{
\item{type}{Reference-list type.}
}
\value{
Character scalar describing the reference-list type.
}
\description{
Reference-list help text
}
