% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation.R
\name{search_exact}
\alias{search_exact}
\title{Exact match search via CompToxR bulk API}
\usage{
search_exact(unique_names)
}
\arguments{
\item{unique_names}{Character vector of unique chemical names}
}
\value{
Tibble with searchValue, dtxsid, preferredName, searchName, rank
}
\description{
Exact match search via CompToxR bulk API
}
