% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_pipeline.R
\docType{data}
\name{CHIRAL_PLACEHOLDER_PREFIX}
\alias{CHIRAL_PLACEHOLDER_PREFIX}
\title{Prefix for chiral designation placeholders}
\format{
An object of class \code{character} of length 1.
}
\usage{
CHIRAL_PLACEHOLDER_PREFIX
}
\description{
Used by protect_chiral_designations() to replace chiral markers before
downstream enclosure stripping (Step 6a) removes them.
}
\keyword{datasets}
