% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_harmonizer.R
\name{is_molarity_unit}
\alias{is_molarity_unit}
\title{Check if a unit is a molarity unit}
\usage{
is_molarity_unit(unit)
}
\arguments{
\item{unit}{Character vector of unit strings (normalized)}
}
\value{
Logical vector
}
\description{
Check if a unit is a molarity unit
}
\keyword{internal}
