% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation.R
\name{validate_and_lookup_cas}
\alias{validate_and_lookup_cas}
\title{Validate CAS numbers and lookup DTXSID for valid ones}
\usage{
validate_and_lookup_cas(unique_cas)
}
\arguments{
\item{unique_cas}{Character vector of CAS-like strings}
}
\value{
Tibble with original_cas, validated_cas, is_valid, dtxsid, preferredName
}
\description{
Validate CAS numbers and lookup DTXSID for valid ones
}
